filter out specific directories from creating autoloads
Note that this is a fairly crude method that removes whole swaths of
elisp code in folders, which can cause dependency issues with
individual elisp files (see eieio-datadebug.el or iswitchb.el)

It also requires manually syncing the ignored directories in
'build-aux/update-subdirs' and 'lisp/Makefile.in'. A better
solution (maybe in the future) would be to statically determine the
dependency tree of all the elisp files and prune definitions that way.